feature create promotions, apply discount code for tour and filter promotion by date

+++ WebTour/src/main/java/WebApplication/WebTour/Model/Paymentmethod.java
@@ -20,7 +20,7 @@ public class Paymentmethod implements Serializable{
 	@Id
  	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "paymethod_id")
-    private int paymethodId ;
+    private long paymethodId ;
 
     @Column(name = "name")
     private String name;
@@ -44,7 +44,7 @@ public class Paymentmethod implements Serializable{
 		status = status;
 	}
 
-	public int getPaymethodId() {
+	public long getPaymethodId() {
 		return paymethodId;
 	}
 }
